requirements.txt
numpy==1.21.6
openai==0.27.4
pandas==1.4.4
plotly==5.9.0
python-dotenv==1.0.0
scikit_learn==1.2.2
tqdm==4.64.1
